feat: implement altimeter

+++ unity_project/unity_sim/Assets/Simulation/ROSBridge/GeometryMsgs/Vector3.cs
@@ -2,20 +2,19 @@ using Newtonsoft.Json;
 
 namespace ROSBridge.GeometryMsgs
 {
-
     public struct Vector3
     {
         [JsonIgnore]
         public static readonly string ROSMessageType = "geometry_msgs/msg/Vector3";
 
         [JsonProperty("x")]
-        public float X { get; set; }
+        public double X { get; set; }
 
         [JsonProperty("y")]
-        public float Y { get; set; }
+        public double Y { get; set; }
 
         [JsonProperty("z")]
-        public float Z { get; set; }
+        public double Z { get; set; }
     }
 
 }
